change: BigInteger money changed to double

+++ src/main/java/me/jazzy/opos/dto/PizzaDto.java
@@ -3,13 +3,11 @@ package me.jazzy.opos.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 @AllArgsConstructor
 public class PizzaDto {
     private String name;
-    private BigInteger money;
+    private double money;
     private Long ingredientId;
     private Long categoryId;
 }
